Address review feedback: spelling, file limits, wrapper types, execplan outcomes
- Fix en-GB-oxendict -ize spellings ('recognises' -> 'recognizes') in
  docs/cargo-orthohelp-design.md and docs/design.md
- Update execplan Outcomes section with completed milestone summary
- Extract reject_subcommand_ortho_config_attrs into clap_attrs.rs to
  bring parse/mod.rs under 400 lines
- Add unwrap_known_wrapper() to handle Option<T>/Vec<T> wrappers around
  subcommand enum types in build_subcommands_metadata
- Extract subcommand tests and fixtures into docs_ir_subcommands.rs to
  bring docs_ir.rs under 400 lines
