+++ docs/design.md
@@ -1036,7 +1036,7 @@ generated documentation, generated agent context, and enforceable CLI policy.
 - **Generate recursive subcommand documentation metadata (2026-05-24):**
   Introduce the `OrthoConfigSubcommandDocs` companion trait and derive for
   `clap::Subcommand` enums, as accepted in ADR-004. `OrthoConfig` now
-  recognises a `#[command(subcommand)]` selector field, skips it as a
+  recognizes a `#[command(subcommand)]` selector field, skips it as a
   configuration field, and fills `DocMetadata.subcommands` from the enum's
   per-variant metadata in declaration order.
